rename: 게시글 관련 api, 타입 파일명 변경
