+++ src/api/post/index.ts
@@ -1,5 +1,5 @@
 import http from '@/api/core';
-import { CreatePost } from '@/types/CreatePost';
+import { CreatePost } from '@/types/post';
 
 export const createPost = async (data: CreatePost) => {
   await http.post({
